pyproject.toml
[project]
name = "konkani-amr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai",
    "python-dotenv",
    "ruff",
    "tqdm>=4.67.3",
]
